CDAP-15860 Facebook Ads (SaaS source) - batch source - breakdown to base config
