+++ src/main/java/io/cdap/plugin/facebook/source/batch/FacebookBatchSourceConfig.java
@@ -16,16 +16,12 @@
 
 package io.cdap.plugin.facebook.source.batch;
 
-import com.facebook.ads.sdk.AdsInsights;
 import com.google.common.base.Strings;
 import io.cdap.cdap.api.annotation.Description;
 import io.cdap.cdap.api.annotation.Macro;
 import io.cdap.cdap.api.annotation.Name;
 import io.cdap.cdap.etl.api.FailureCollector;
 import io.cdap.plugin.facebook.source.common.config.BaseSourceConfig;
-import io.cdap.plugin.facebook.source.common.config.Breakdowns;
-import io.cdap.plugin.facebook.source.common.config.SourceConfigHelper;
-import io.cdap.plugin.facebook.source.common.exceptions.IllegalBreakdownException;
 
 import javax.annotation.Nullable;
 
@@ -33,22 +29,8 @@ import javax.annotation.Nullable;
  * Provides all required configuration for reading Facebook Insights.
  */
 public class FacebookBatchSourceConfig extends BaseSourceConfig {
-  public static final String PROPERTY_BREAKDOWN = "breakdown";
-  public static final String PROPERTY_ADDITIONAL_BREAKDOWN = "additionalBreakdown";
   public static final String PROPERTY_SORTING = "sorting";
   public static final String PROPERTY_SORT_DIRECTION = "sortDirection";
-
-  @Name(PROPERTY_BREAKDOWN)
-  @Description("Primary breakdown.")
-  @Nullable
-  @Macro
-  protected String breakdown;
-
-  @Name(PROPERTY_ADDITIONAL_BREAKDOWN)
-  @Description("Additional breakdown. Can be selected with primary breakdowns marked by '*'.")
-  @Nullable
-  @Macro
-  protected String additionalBreakdown;
 
   @Name(PROPERTY_SORTING)
   @Description("Field name to sort results by.")
@@ -66,22 +48,6 @@ public class FacebookBatchSourceConfig extends BaseSourceConfig {
     super(referenceName);
   }
 
-  public Breakdowns getBreakdown() {
-    if (!Strings.isNullOrEmpty(breakdown) && !"none".equals(breakdown)) {
-      Breakdowns result = SourceConfigHelper.parseBreakdowns(breakdown);
-      if (!Strings.isNullOrEmpty(additionalBreakdown) && !"none".equals(additionalBreakdown)) {
-        AdsInsights.EnumActionBreakdowns additionalActionBreakdown =
-          SourceConfigHelper.actionBreakdownFromString(additionalBreakdown);
-        if (result.isJoinableWithAction() && result.getActionBreakdowns().contains(additionalActionBreakdown)) {
-          result.getActionBreakdowns().add(additionalActionBreakdown);
-        }
-      }
-      return result;
-    } else {
-      return null;
-    }
-  }
-
   @Nullable
   public String getSorting() {
     if (!Strings.isNullOrEmpty(sorting) && !Strings.isNullOrEmpty(sortDirection)) {
@@ -95,20 +61,7 @@ public class FacebookBatchSourceConfig extends BaseSourceConfig {
   public void validate(FailureCollector failureCollector) {
     super.validate(failureCollector);
 
-    validateBreakdowns(failureCollector);
     validateSorting(failureCollector);
-  }
-
-  void validateBreakdowns(FailureCollector failureCollector) {
-    if (!containsMacro(PROPERTY_BREAKDOWN)) {
-      try {
-        getBreakdown();
-      } catch (IllegalBreakdownException ex) {
-        failureCollector
-          .addFailure(ex.getMessage(), "Fix invalid breakdown value.")
-          .withConfigProperty(PROPERTY_BREAKDOWN);
-      }
-    }
   }
 
   void validateSorting(FailureCollector failureCollector) {
